pyproject.toml
[project]
name = "arcanomy-motion"
version = "0.1.0"
description = "AI-powered short-form video production pipeline for financial content"
requires-python = ">=3.10"
dependencies = [
    # LLM Providers
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "google-generativeai>=0.8.0",
    
    # Voice Synthesis
    "elevenlabs>=1.0.0",
    
    # Core utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    
    # Media processing
    "pillow>=10.0.0",
    "requests>=2.31.0",
    
    # Auth (for API clients)
    "pyjwt>=2.8.0",
]

[project.scripts]
arcanomy = "src.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
